아마존 파일 다운로드 및 삭제시 존재하지 않는 파일에 대한 에러처리

+++ blockboard/src/main/java/com/board/project/blockboard/service/AmazonS3Service.java
@@ -26,10 +26,12 @@ import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.amazonaws.services.s3.model.S3Object;
 import com.amazonaws.services.s3.model.S3ObjectInputStream;
 import com.board.project.blockboard.common.constant.ConstantData;
+import com.board.project.blockboard.common.exception.FileValidException;
 import com.board.project.blockboard.key.Key;
 import java.io.File;
 import java.io.InputStream;
 import java.util.List;
+import javax.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.tomcat.util.bcel.Const;
 import org.springframework.stereotype.Service;
@@ -66,15 +68,22 @@ public class AmazonS3Service {
     return null;
   }
 
-  public S3ObjectInputStream download(String fileName, String bucket) {
+  public S3ObjectInputStream download(String fileName, String bucket, HttpServletResponse response) {
     if (amazonS3 != null) {
+
       try {
+        if(!amazonS3.doesObjectExist(bucket,fileName)){ //버킷에 파일이 존재하지 않을때.
+          throw new FileValidException();
+        }
         S3Object o = amazonS3.getObject(ConstantData.BUCKET_FILE, fileName);
         S3ObjectInputStream s3is = o.getObjectContent();
         return s3is;
       } catch (AmazonServiceException e) {
         System.err.println(e.getErrorMessage());
         System.exit(1);
+      } catch (FileValidException fve) {
+        fve.sendError(response,"존재하지 않는 파일입니다.");
+        fve.printStackTrace();
       } finally {
         amazonS3 = null;
       }
@@ -82,10 +91,12 @@ public class AmazonS3Service {
     return null;
   }
 
-  public boolean deleteFile(String fileName, String bucket) {
+  public boolean deleteFile(String fileName, String bucket,HttpServletResponse response) {
     if (amazonS3 != null) {
       try {
-
+        if(!amazonS3.doesObjectExist(bucket,fileName)){ //버킷에 파일이 존재하지 않을때.
+          throw new FileValidException();
+        }
         amazonS3.deleteObject(ConstantData.BUCKET_FILE, fileName);
         if (amazonS3.doesObjectExist(ConstantData.BUCKET_FILE, fileName)) {
           return false;
@@ -95,6 +106,9 @@ public class AmazonS3Service {
         System.err.println(e.getErrorMessage());
         System.exit(1);
         return false;
+      } catch (FileValidException fve) {
+        fve.sendError(response,"이미 존재하지 않는 파일입니다.");
+        fve.printStackTrace();
       } finally {
         amazonS3 = null;
       }
